iptables: wait for xtables lock
Normally when the xtables lock is held by another process iptables
aborts. When containers are starting on a new (or restarting) node the
lock is very busy leading to frequent failures. Instead, wait for the
lock up to 5 seconds, retrying every 100ms; these are the default values
used by kube-proxy.